Unify RS tests into a single file
